+++ src/ast_formatter/printer/key_val.rs
@@ -1,16 +1,18 @@
 use super::{
-    assignment_operator::AssignmentOperator, comment::Comment, operator::Operator, ASTPrint,
+    assignment_operator::AssignmentOperator,
+    comment::Comment,
+    indices::{ArrayIndex, Index},
+    operator::Operator,
+    ASTPrint,
 };
 
-#[derive(Debug)]
+#[derive(Debug, Default)]
 pub struct KeyVal {
     pub operator: Option<Operator>,
     pub key: String,
     pub needs: Option<String>,
-    // TODO: Replace with an enum
-    pub index: Option<String>,
-    // TODO: Replace with a struct
-    pub array_index: Option<String>,
+    pub index: Option<Index>,
+    pub array_index: Option<ArrayIndex>,
     pub assignment_operator: AssignmentOperator,
     pub val: String,
     pub comment: Option<Comment>,
@@ -25,8 +27,10 @@ impl ASTPrint for KeyVal {
             self.operator.clone().unwrap_or_default(),
             self.key,
             self.needs.clone().unwrap_or_default(),
-            self.index.clone().unwrap_or_default(),
-            self.array_index.clone().unwrap_or_default(),
+            self.index.clone().map_or("".to_owned(), |i| i.to_string()),
+            self.array_index
+                .clone()
+                .map_or("".to_owned(), |i| i.to_string()),
             self.assignment_operator,
             self.val,
             self.comment.clone().unwrap_or_default(),
