+++ tests/weird_configs/weird_index_selection.cfg
@@ -1,1 +1,3 @@
+NODE { @key[*] *= 2 }
+
 NODE { @key,*[*, ] *= 2 }
